+++ src/frontend/tests/core/integrations/Custom Component Generator.spec.ts
@@ -8,6 +8,7 @@ import { getAllResponseMessage } from "../../utils/get-all-response-message";
 import { removeOldApiKeys } from "../../utils/remove-old-api-keys";
 import { selectGptModel } from "../../utils/select-gpt-model";
 import { updateOldComponents } from "../../utils/update-old-components";
+import { waitForOpenModalWithChatInput } from "../../utils/wait-for-open-modal";
 
 test(
   "Custom Component Generator",
@@ -61,9 +62,7 @@ test(
       .last()
       .isVisible();
 
-    await page.waitForSelector('[data-testid="input-chat-playground"]', {
-      timeout: 100000,
-    });
+    await waitForOpenModalWithChatInput(page);
 
     const textContents = await getAllResponseMessage(page);
     expect(textContents.length).toBeGreaterThan(100);
