+++ src/frontend/tests/core/integrations/Prompt Chaining.spec.ts
@@ -1,6 +1,14 @@
-import { test } from "@playwright/test";
+import { expect, test } from "@playwright/test";
 import * as dotenv from "dotenv";
 import path from "path";
+import { addNewApiKeys } from "../../utils/add-new-api-keys";
+import { adjustScreenView } from "../../utils/adjust-screen-view";
+import { awaitBootstrapTest } from "../../utils/await-bootstrap-test";
+import { getAllResponseMessage } from "../../utils/get-all-response-message";
+import { removeOldApiKeys } from "../../utils/remove-old-api-keys";
+import { selectGptModel } from "../../utils/select-gpt-model";
+import { updateOldComponents } from "../../utils/update-old-components";
+import { waitForOpenModalWithChatInput } from "../../utils/wait-for-open-modal";
 
 test(
   "Prompt Chaining",
@@ -16,68 +24,20 @@ test(
     }
 
     await page.goto("/");
-    await page.waitForSelector('[data-testid="mainpage_title"]', {
-      timeout: 30000,
-    });
-
-    await page.waitForSelector('[id="new-project-btn"]', {
-      timeout: 30000,
-    });
-
-    let modalCount = 0;
-    try {
-      const modalTitleElement = await page?.getByTestId("modal-title");
-      if (modalTitleElement) {
-        modalCount = await modalTitleElement.count();
-      }
-    } catch (error) {
-      modalCount = 0;
-    }
-
-    while (modalCount === 0) {
-      await page.getByText("New Flow", { exact: true }).click();
-      await page.waitForSelector('[data-testid="modal-title"]', {
-        timeout: 3000,
-      });
-      modalCount = await page.getByTestId("modal-title")?.count();
-    }
+    await awaitBootstrapTest(page);
 
     await page.getByTestId("side_nav_options_all-templates").click();
-    await page.getByRole("heading", { name: "Basic Prompting" }).click();
+    await page.getByRole("heading", { name: "Prompt Chaining" }).click();
 
     await page.waitForSelector('[data-testid="fit_view"]', {
       timeout: 100000,
     });
 
-    await page.getByTestId("fit_view").click();
-    await page.getByTestId("zoom_out").click();
-    await page.getByTestId("zoom_out").click();
-    await page.getByTestId("zoom_out").click();
-
-    let outdatedComponents = await page
-      .getByTestId("icon-AlertTriangle")
-      .count();
-
-    while (outdatedComponents > 0) {
-      await page.getByTestId("icon-AlertTriangle").first().click();
-      outdatedComponents = await page.getByTestId("icon-AlertTriangle").count();
-    }
-
-    let filledApiKey = await page.getByTestId("remove-icon-badge").count();
-    while (filledApiKey > 0) {
-      await page.getByTestId("remove-icon-badge").first().click();
-      filledApiKey = await page.getByTestId("remove-icon-badge").count();
-    }
-
-    const apiKeyInput = page.getByTestId("popover-anchor-input-api_key");
-    const isApiKeyInputVisible = await apiKeyInput.isVisible();
-
-    if (isApiKeyInputVisible) {
-      await apiKeyInput.fill(process.env.OPENAI_API_KEY ?? "");
-    }
-
-    await page.getByTestId("dropdown_str_model_name").click();
-    await page.getByTestId("gpt-4o-1-option").click();
+    await adjustScreenView(page);
+    await updateOldComponents(page);
+    await removeOldApiKeys(page);
+    await addNewApiKeys(page);
+    await selectGptModel(page);
 
     await page.getByTestId("button_run_chat output").click();
     await page.waitForSelector("text=built successfully", { timeout: 30000 });
@@ -92,42 +52,9 @@ test(
       .last()
       .isVisible();
 
-    await page.waitForSelector('[data-testid="input-chat-playground"]', {
-      timeout: 100000,
-    });
+    await waitForOpenModalWithChatInput(page);
 
-    await page
-      .getByTestId("input-chat-playground")
-      .last()
-      .fill("Say hello as a pirate");
-
-    await page.waitForSelector('[data-testid="button-send"]', {
-      timeout: 100000,
-    });
-
-    await page.getByTestId("button-send").last().click();
-
-    await page.waitForSelector("text=matey", {
-      timeout: 100000,
-    });
-
-    await page.getByText("matey").last().isVisible();
-    await page.getByText("Default Session").last().click();
-
-    await page.getByText("timestamp", { exact: true }).last().isVisible();
-    await page.getByText("text", { exact: true }).last().isVisible();
-    await page.getByText("sender", { exact: true }).last().isVisible();
-    await page.getByText("sender_name", { exact: true }).last().isVisible();
-    await page.getByText("session_id", { exact: true }).last().isVisible();
-    await page.getByText("files", { exact: true }).last().isVisible();
-
-    await page.getByRole("gridcell").last().isVisible();
-    await page.getByRole("combobox").click();
-    await page.getByLabel("Delete").click();
-    await page.waitForSelector('[data-testid="input-chat-playground"]', {
-      timeout: 100000,
-    });
-
-    await page.getByTestId("input-chat-playground").last().isVisible();
+    const textContents = await getAllResponseMessage(page);
+    expect(textContents.length).toBeGreaterThan(100);
   },
 );
